requirements.txt
colorama==0.4.1
easygui==0.98.1
mysql-connector-python==8.0.22
prettytable==1.0.1
pyfiglet==0.8.post1
PyPDF2==1.26.0
pyttsx3==2.7
termcolor==1.1.0
tqdm==4.54.1
pywin32
